Add user roles, security method and access scopes (#18)
* complete authentication middleware logic

* add user roles, security method and access scopes

* fix some typos

* add role entity, fix minor issues, mentioned in PR comments

* add OneToMany User - Role relationship

* fix namings

+++ src/test-support/database-helpers.ts
@@ -24,8 +24,10 @@ export function useTestDatabase(): void {
 async function clearDatabaseTables(connection: Connection): Promise<void> {
     await Promise.all(
         connection.entityMetadatas.map(async (entity) => {
-            const repository = connection.getRepository(entity.name);
-            await repository.query(`TRUNCATE TABLE "${entity.tableName}" RESTART IDENTITY CASCADE`);
+            if (entity.name !== 'Role') {
+                const repository = connection.getRepository(entity.name);
+                await repository.query(`TRUNCATE TABLE "${entity.tableName}" RESTART IDENTITY CASCADE`);
+            }
         })
     );
 }
